Refactor game startup with simple engine
