+++ main.js
@@ -1,19 +1,11 @@
 // main.js
-import { Game } from './src/game.js';
+import { GameInitializer } from './src/core/GameInitializer.js';
 import { registerServiceWorker } from './src/utils/swRegister.js';
-
-let game = null;
-
-function initializeAudio() {
-    if (game) {
-        game.startBGM();
-    }
-}
 
 window.onload = () => {
     registerServiceWorker();
-    game = new Game();
-    game.start();
-    document.addEventListener('keydown', initializeAudio, { once: true });
-    document.addEventListener('click', initializeAudio, { once: true });
+    const canvas = document.getElementById('battleCanvas');
+    const context = canvas.getContext('2d');
+    const initializer = new GameInitializer(context);
+    initializer.start();
 };
