+++ src/core/GameInitializer.js
@@ -1,8 +1,9 @@
 import { AssetLoader } from '../assetLoader.js';
+import { setupAndStartGame } from '../game.js';
 
 export class GameInitializer {
-  constructor(game) {
-    this.game = game;
+  constructor(context) {
+    this.context = context;
     this.loader = new AssetLoader();
   }
 
@@ -54,8 +55,9 @@ export class GameInitializer {
     l.loadEmblemImages();
     l.loadVfxImages();
 
-    l.onReady((assets) => {
-      this.game.init(assets);
+    l.onReady(() => {
+      console.log('에셋 로딩 완료, 게임 설정 시작...');
+      setupAndStartGame(this.context);
     });
   }
 }
