+++ src/inputManager.js
@@ -1,0 +1,15 @@
+export class InputManager {
+    constructor() {
+        this.keys = {};
+        window.addEventListener('keydown', e => {
+            this.keys[e.key.toLowerCase()] = true;
+        });
+        window.addEventListener('keyup', e => {
+            this.keys[e.key.toLowerCase()] = false;
+        });
+    }
+
+    isPressed(key) {
+        return this.keys[key] || false;
+    }
+}
